Fixup file reading in built product
  *Worked with the file reader so that it would work during the
development process and in the deployed application. Mostly involved
appending __dirname and using the path.join() method to stitch together
the path to files in an agnostic fashion.
  *Got the database connectivity working in the deployed application.
This includes the application of the migration files to the app once
it's deployed, each time it's opened.
  *Upgraded to Electron 5.
  *Added logging, and production persistence of the database object to
the userData folder.
  *Simplified the menu at the top and got rid of extraneous pieces.
  *Removed IPC_EVENT inclusion. Will have to figure out how to do this
later. Apparently the application is not currently configured to include
web-style typescript compiled file into main process written code.
  *Completed the non-functional GUI for the 'add bunny' task. Also
started using momentjs to populate the dates and datepicker of that UI
interface.
  *Migrated the UI over from bootstrap to angular's material-UI 2.